# Optimize transform