--- a/transform/rotate.go
+++ b/transform/rotate.go
@@ -31,10 +31,15 @@ func Rotate(img image.Image, angle float64, options *RotationOptions) *image.RGB
 	src := clone.AsRGBA(img)
 	srcW, srcH := src.Bounds().Dx(), src.Bounds().Dy()
 
-	// Return early if nothing to do
+	supersample := false
 	absAngle := int(math.Abs(angle) + 0.5)
 	if absAngle%360 == 0 {
+		// Return early if nothing to do
 		return src
+	} else if absAngle%90 != 0 {
+		// Supersampling is required for non-special angles
+		// Special angles = 90, 180, 270,...
+		supersample = true
 	}
 
 	// Config defaults
@@ -49,10 +54,12 @@ func Rotate(img image.Image, angle float64, options *RotationOptions) *image.RGB
 		}
 	}
 
-	// Supersample, currently hard set to 2x
-	srcW, srcH = srcW*2, srcH*2
-	src = Resize(src, srcW, srcH, NearestNeighbor)
-	pivotX, pivotY = pivotX*2, pivotY*2
+	if supersample {
+		// Supersample, currently hard set to 2x
+		srcW, srcH = srcW*2, srcH*2
+		src = Resize(src, srcW, srcH, NearestNeighbor)
+		pivotX, pivotY = pivotX*2, pivotY*2
+	}
 
 	// Convert to radians, positive degree maps to clockwise rotation
 	angleRadians := -angle * (math.Pi / 180)
@@ -110,8 +117,10 @@ func Rotate(img image.Image, angle float64, options *RotationOptions) *image.RGB
 		}
 	})
 
-	// Downsample to original bounds as part of the Supersampling
-	dst = Resize(dst, dstW/2, dstH/2, Linear)
+	if supersample {
+		// Downsample to original bounds as part of the Supersampling
+		dst = Resize(dst, dstW/2, dstH/2, Linear)
+	}
 
 	return dst
 }